Add segments that give greater detail about vcs
Note: In the current implementation, I have only added git support.
This is because I generally do not deal with svn and hg repositories
at the moment and do not know the best way to implement them. In some
cases, these segments may be meaningless for certain vcs's (i.e.
staged/prepared files may not be meaningful in svn). A description is
given below

* new file:   segments/vcs_compare.sh
    This compares the current branch with that of it's remote branch (as
    reached with "git fetch"). It may look like: ↓ 3 ↑ 2
* new file:   segments/vcs_modified.sh
    This shows the number of items that are modified and not staged
* new file:   segments/vcs_others.sh
    This shows the number of items that are new / un-tracked by the vcs
* new file:   segments/vcs_staged.sh
    These are items that are staged / prepared to be commited into the vcs
* modified:   status-left.sh
    These are set to on by default