Feat: Add a base16 default theme
This adds a theme based on the default/bubble that
uses named tmux colors (the base16 palette).

This allows for theme colors to be inherited from
the terminal/shell, which is useful for things like
VS Code color integration as well as other terminal
with built in themes.

It also reads from a TMUX_POWERLINE_BUBBLE_SEPARATORS environment
variable to toggle angled vs rounded separators.